move normalize functions into column_utils.py
