+++ python/cuspatial/cuspatial/tests/test_trajectory.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2019, NVIDIA CORPORATION.
+# Copyright (c) 2019-2020, NVIDIA CORPORATION.
 
 import numpy as np
 import pytest
