+++ python/cuspatial/cuspatial/utils/trajectory_utils.py
@@ -1,45 +1,4 @@
-import numpy as np
-
-from cudf.utils.dtypes import is_datetime_dtype
-
-
-def normalize_point_columns(xs, ys):
-    """
-    Normalize the input columns by inferring a common floating point dtype.
-
-    If the common dtype isn't a floating point dtype, promote the common dtype
-    to float64.
-
-    Parameters
-    ----------
-    {params}
-
-    Returns
-    -------
-    tuple : the input columns cast to the inferred common floating point dtype
-    """
-    dtype = np.result_type(xs.dtype, ys.dtype)
-    if not np.issubdtype(dtype, np.floating):
-        dtype = np.float32 if dtype.itemsize <= 4 else np.float64
-    return xs.astype(dtype), ys.astype(dtype)
-
-
-def normalize_timestamp_column(ts, fallback_dtype="datetime64[ms]"):
-    """
-    Normalize the input timestamp column to one of the cuDF timestamp dtypes.
-
-    If the input column's dtype isn't an np.datetime64, cast the column to the
-    supplied `fallback_dtype` parameter.
-
-    Parameters
-    ----------
-    {params}
-
-    Returns
-    -------
-    column : the input column
-    """
-    return ts if is_datetime_dtype(ts.dtype) else ts.astype(fallback_dtype)
+# Copyright (c) 2020, NVIDIA CORPORATION.
 
 
 def get_ts_struct(ts):
